requirements.txt
transformers==4.30.0
datasets==2.13.0
gradio==3.34.0
accelerate==0.20.3
torch==2.0.1
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
tqdm==4.65.0
regex==2023.6.3
pillow==9.5.0
jinja2==3.1.2
diffusers==0.19.3
ftfy==6.1.1
